Move setup_proxy into main __init__ file
f55af10f1767d39fdba65fb4c17beee526f96748 should have gone into the main
__init__.py, not the one in ./lib, since this is where we are keeping this
general setup material at the moment. This moves the function to the appropriate
location.
