+++ __init__.py
@@ -93,6 +93,18 @@ def ENGINE_LIST():
     return engine_list
 
 
+def set_proxy():
+  proxies = ["https_proxy", "http_proxy", "ftp_proxy", "HTTP_PROXY", "HTTPS_PROXY", "FTP_PROXY"]
+  for proxy in proxies:
+    if os.getenv(proxy):
+      if len(os.environ[proxy]) != 0:
+        for i in proxies:
+          os.environ[i] = os.environ[proxy]
+        break
+
+set_proxy()
+
+
 sample_script = """# basic information about the script
 name: Mammal Life History Database - Ernest, et al., 2003
 shortname: MammalLH
